pyproject.toml
[tool.poetry]
name = "ethjsonrpc"
version = "0.1.0"
description = ""
authors = ["Clément Walter <clement0walter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
cairo-lang = "^0.10.1"
starknet-py = "0.14.0a0"
fastapi = "^0.94.1"
pydantic = "^1.10.6"
uvicorn = "^0.21.0"
python-dotenv = "^1.0.0"
typeguard = "2.13.0"
rlp = "<=3"
py-evm = "^0.6.1a2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
autoflake = "^2.0.2"
ipython = "^8.11.0"
ipykernel = "^6.21.3"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
